+++ e2e/commands/untag.e2e.js
@@ -128,6 +128,19 @@ describe('bit untag command', function () {
         });
       });
     });
+    describe('when tagging non-existing component', () => {
+      let output;
+      before(() => {
+        try {
+          helper.runCmd('bit untag non-exist-scope/non-exist-comp');
+        } catch (err) {
+          output = err.message;
+        }
+      });
+      it('should show an descriptive error', () => {
+        expect(output).to.have.string('fatal: component with id "non-exist-scope/non-exist-comp" was not found');
+      });
+    });
   });
   describe('untag multiple components (--all flag)', () => {
     let localScope;
@@ -241,6 +254,21 @@ describe('bit untag command', function () {
             'unable to untag utils/is-type@0.0.1, the version 0.0.1 has the following dependent(s)'
           );
         });
+      });
+    });
+    describe('untag only the dependent', () => {
+      let untagOutput;
+      before(() => {
+        helper.getClonedLocalScope(localScope);
+        untagOutput = helper.runCmd('bit untag utils/is-string');
+      });
+      it('should untag successfully the dependent', () => {
+        expect(untagOutput).to.have.string('1 component(s) were untagged');
+        expect(untagOutput).to.have.string('utils/is-string');
+      });
+      it('should leave the dependency intact', () => {
+        const output = helper.listLocalScope();
+        expect(output).to.have.string('utils/is-type');
       });
     });
   });