+++ e2e/fixtures/components/vue/directives/autofocus.js
@@ -1,0 +1,7 @@
+export default {
+  inserted(el, { value }) {
+    if (value) {
+      el.focus();
+    }
+  }
+};