+++ src/cli/commands/public-cmds/clear-cache-cmd.js
@@ -1,9 +1,10 @@
 /** @flow */
 import chalk from 'chalk';
 import fs from 'fs-extra';
+import roadRunner from 'roadrunner';
 import Command from '../../command';
 
-const { MODULES_CACHE_DIR } = require('../../../constants');
+const { MODULES_CACHE_DIR, MODULES_CACHE_FILENAME, BIT_VERSION } = require('../../../constants');
 
 export default class ClearCache extends Command {
   name = 'clear-cache';
@@ -15,6 +16,8 @@ export default class ClearCache extends Command {
   action(): Promise<any> {
     fs.removeSync(MODULES_CACHE_DIR);
     fs.mkdirsSync(MODULES_CACHE_DIR);
+    roadRunner.reset(MODULES_CACHE_FILENAME);
+    roadRunner.set('CACHE_BREAKER', { version: BIT_VERSION });
     return Promise.resolve();
   }