+++ src/consumer/component/consumer-component.js
@@ -97,6 +97,9 @@ export default class Component {
   dependenciesSavedAsComponents: ?boolean = true; // otherwise they're saved as npm packages
   originallySharedDir: ?PathLinux; // needed to reduce a potentially long path that was used by the author
   _wasOriginallySharedDirStripped: ?boolean; // whether stripOriginallySharedDir() method had been called, we don't want to strip it twice
+  loadedFromFileSystem: boolean = false; // whether a component was loaded from the filesystem or converted from the model
+  componentMap: ?ComponentMap; // always populated when the loadedFromFileSystem is true
+  componentFromModel: ?Component; // populated when loadedFromFileSystem is true and it exists in the model
   isolatedEnvironment: IsolatedEnvironment;
   missingDependencies: ?Object;
   deprecated: boolean;
@@ -840,8 +843,26 @@ export default class Component {
   }
 
   async toComponentWithDependencies(consumer: Consumer): Promise<ComponentWithDependencies> {
-    const dependencies = await this.dependencies.getDependenciesComponents(consumer, this);
-    const devDependencies = await this.devDependencies.getDependenciesComponents(consumer, this);
+    const getFlatten = (dev: boolean = false) => {
+      const field = dev ? 'flattenedDevDependencies' : 'flattenedDependencies';
+      // when loaded from filesystem, it doesn't have the flatten, fetch them from model.
+      return this.loadedFromFileSystem ? this.componentFromModel[field] : this[field];
+    };
+    const getDependenciesComponents = (ids: BitIds) => {
+      return Promise.all(
+        ids.map((dependencyId) => {
+          if (consumer.bitMap.isExistWithSameVersion(dependencyId)) {
+            return consumer.loadComponent(dependencyId);
+          }
+          // when dependencies are imported as npm packages, they are not in bit.map
+          this.dependenciesSavedAsComponents = false;
+          return consumer.scope.loadComponent(dependencyId, false);
+        })
+      );
+    };
+
+    const dependencies = await getDependenciesComponents(getFlatten());
+    const devDependencies = await getDependenciesComponents(getFlatten(true));
     return new ComponentWithDependencies({ component: this, dependencies, devDependencies });
   }
 
@@ -910,13 +931,13 @@ export default class Component {
     componentMap: ComponentMap,
     id: BitId,
     consumer: Consumer,
-    componentFromModel: ModelComponent
+    componentFromModel: Component
   }): Component {
     const consumerPath = consumer.getPath();
     const consumerBitJson: ConsumerBitJson = consumer.bitJson;
     const bitMap: BitMap = consumer.bitMap;
-    const deprecated = componentFromModel ? componentFromModel.component.deprecated : false;
-    let dists = componentFromModel ? componentFromModel.component.dists.get() : undefined;
+    const deprecated = componentFromModel ? componentFromModel.deprecated : false;
+    let dists = componentFromModel ? componentFromModel.dists.get() : undefined;
     let packageDependencies;
     let devPackageDependencies;
     let peerPackageDependencies;
@@ -947,7 +968,7 @@ export default class Component {
 
       return sourceFiles;
     };
-    if (!fs.existsSync(bitDir)) return Promise.reject(new ComponentNotFoundInPath(bitDir));
+    if (!fs.existsSync(bitDir)) throw new ComponentNotFoundInPath(bitDir);
     const files = componentMap.files;
     // Load the base entry from the root dir in map file in case it was imported using -path
     // Or created using bit create so we don't want all the path but only the relative one
@@ -965,7 +986,7 @@ export default class Component {
     // by default, imported components are not written with bit.json file.
     // use the component from the model to get their bit.json values
     if (!fs.existsSync(path.join(bitDir, BIT_JSON)) && componentFromModel) {
-      bitJson.mergeWithComponentData(componentFromModel.component);
+      bitJson.mergeWithComponentData(componentFromModel);
     }
 
     // Remove dists if compiler has been deleted