+++ src/scope/scope.js
@@ -22,9 +22,7 @@ import {
   BITS_DIRNAME,
   BIT_VERSION,
   DEFAULT_BIT_VERSION,
-  LATEST_BIT_VERSION,
-  BIT_GIT_DIR,
-  DOT_GIT_DIR
+  LATEST_BIT_VERSION
 } from '../constants';
 import { ScopeJson, getPath as getScopeJsonPath } from './scope-json';
 import {
@@ -44,7 +42,6 @@ import { Repository, Ref, BitObject } from './objects';
 import ComponentWithDependencies from './component-dependencies';
 import VersionDependencies from './version-dependencies';
 import SourcesRepository from './repositories/sources';
-import npmClient from '../npm-client';
 import Consumer from '../consumer/consumer';
 import { index } from '../search/indexer';
 import loader from '../cli/loader';
@@ -67,7 +64,7 @@ import Component from '../consumer/component/consumer-component';
 import DependencyGraph from './graph/graph';
 
 const removeNils = R.reject(R.isNil);
-const pathHasScope = pathHas([OBJECTS_DIR, BIT_HIDDEN_DIR, pathLib.join(DOT_GIT_DIR, BIT_GIT_DIR)]);
+const pathHasScope = pathHas([OBJECTS_DIR, BIT_HIDDEN_DIR]);
 
 export type ScopeDescriptor = {
   name: string
@@ -768,7 +765,7 @@ export default class Scope {
     });
   }
 
-  async get(id: BitId): Promise<ConsumerComponent> {
+  async get(id: BitId): Promise<ComponentWithDependencies> {
     return this.import(id).then((versionDependencies) => {
       return versionDependencies.toConsumer(this.objects);
     });
@@ -778,7 +775,7 @@ export default class Scope {
    * return a component only when it's stored locally. Don't go to any remote server and don't throw an exception if the
    * component is not there.
    */
-  async getFromLocalIfExist(id: BitId): Promise<?ConsumerComponent> {
+  async getFromLocalIfExist(id: BitId): Promise<?ComponentWithDependencies> {
     const componentFromSources = await this.sources.get(id);
     if (!componentFromSources) return null;
     const versionDependencies = await componentFromSources.toVersionDependencies(id.version, this, this.name);
@@ -1311,91 +1308,6 @@ export default class Scope {
   }
 
   /**
-   * If not specified version, remove all local versions.
-   */
-  async removeLocalVersion(
-    id: BitId,
-    version?: string,
-    force?: boolean = false
-  ): Promise<{ id: BitId, versions: string[] }> {
-    const component: ComponentModel = await this.sources.get(id);
-    const localVersions = component.getLocalVersions();
-    if (!localVersions.length) return Promise.reject(`unable to untag ${id}, the component is not staged`);
-    if (version && !component.hasVersion(version)) {
-      return Promise.reject(`unable to untag ${id}, the version ${version} does not exist`);
-    }
-    if (version && !localVersions.includes(version)) {
-      return Promise.reject(`unable to untag ${id}, the version ${version} was exported already`);
-    }
-    const versionsToRemove = version ? [version] : localVersions;
-
-    if (!force) {
-      const dependencyGraph = await this.getDependencyGraph();
-      versionsToRemove.forEach((versionToRemove) => {
-        const idWithVersion = id.clone();
-        idWithVersion.version = versionToRemove;
-        const dependents = dependencyGraph.getDependentsPerId(idWithVersion);
-        if (dependents.length) {
-          throw new Error(
-            `unable to untag ${id}, the version ${versionToRemove} has the following dependent(s) ${dependents.join(
-              ', '
-            )}`
-          );
-        }
-      });
-    }
-
-    await Promise.all(versionsToRemove.map(ver => this.sources.removeVersion(component, ver, false)));
-    await this.objects.persist();
-
-    if (!component.versionArray.length) {
-      // if all versions were deleted, delete also the component itself from the model
-      await component.remove(this.objects);
-    }
-
-    return { id, versions: versionsToRemove };
-  }
-
-  async removeLocalVersionsForAllComponents(version?: string, force?: boolean = false) {
-    const components = await this.objects.listComponents(false);
-    const candidateComponents = components.filter((component: ComponentModel) => {
-      const localVersions = component.getLocalVersions();
-      if (!localVersions.length) return false;
-      return version ? localVersions.includes(version) : true;
-    });
-    if (!candidateComponents.length) {
-      const versionOutput = version ? `${version} ` : '';
-      return Promise.reject(`No components found with local version ${versionOutput}to untag`);
-    }
-
-    // if no version is given, there is risk of deleting dependencies version without their dependents.
-    if (!force && version) {
-      const dependencyGraph = await this.getDependencyGraph();
-      const candidateComponentsIds = candidateComponents.map((component) => {
-        const bitId = component.toBitId();
-        bitId.version = version;
-        return bitId;
-      });
-      const candidateComponentsIdsStr = candidateComponentsIds.map(id => id.toString());
-
-      candidateComponentsIds.forEach((bitId: BitId) => {
-        const dependents = dependencyGraph.getDependentsPerId(bitId);
-        const dependentsNotCandidates = dependents.filter(dependent => !candidateComponentsIdsStr.includes(dependent));
-        if (dependentsNotCandidates.length) {
-          throw new Error(
-            `unable to untag ${bitId}, the version ${version} has the following dependent(s) ${dependents.join(', ')}`
-          );
-        }
-      });
-    }
-
-    logger.debug(`found ${candidateComponents.length} components to untag`);
-    return Promise.all(
-      candidateComponents.map(component => this.removeLocalVersion(component.toBitId(), version, true))
-    );
-  }
-
-  /**
    * import a component end to end. Including importing the dependencies and installing the npm
    * packages.
    *
@@ -1417,7 +1329,7 @@ export default class Scope {
   }
 
   static load(absPath: string): Promise<Scope> {
-    let scopePath = propogateUntil(absPath, pathHasScope);
+    let scopePath = propogateUntil(absPath);
     if (!scopePath) throw new ScopeNotFound();
     if (fs.existsSync(pathLib.join(scopePath, BIT_HIDDEN_DIR))) {
       scopePath = pathLib.join(scopePath, BIT_HIDDEN_DIR);