+++ src/utils/fs/propogate-until.js
@@ -1,6 +1,8 @@
 /** @flow */
 import * as path from 'path';
 import * as fs from 'fs';
+import findUp from 'find-up';
+import { BIT_GIT_DIR, DOT_GIT_DIR, OBJECTS_DIR, BIT_HIDDEN_DIR } from '../../constants';
 
 function composePath(patternPath: string, patterns: string[]): string[] {
   return patterns.map((pattern) => {
@@ -36,21 +38,11 @@ export function pathHas(patterns: string[]): (absPath: string) => boolean {
  *  // => '/usr/local/var'
  * ```
  */
-export function propogateUntil(fromPath: string, pattern: (path: string) => boolean): ?string {
-  function buildPropogationPaths(): string[] {
-    const paths: string[] = [];
-    const pathParts = fromPath.split(path.sep);
-
-    pathParts.forEach((val, index) => {
-      const part = pathParts.slice(0, index + 1).join('/');
-      if (!part) return;
-      paths.push(part);
-    });
-
-    return paths.reverse();
-  }
-
-  if (pattern(fromPath)) return fromPath;
-  const searchPaths = buildPropogationPaths();
-  return searchPaths.find(searchPath => pattern(searchPath));
+export function propogateUntil(fromPath: string): ?string {
+  if (!fromPath) return null;
+  const filePath = findUp.sync(
+    [OBJECTS_DIR, path.join(BIT_HIDDEN_DIR, OBJECTS_DIR), path.join(DOT_GIT_DIR, BIT_GIT_DIR, OBJECTS_DIR)],
+    { cwd: fromPath }
+  );
+  return path.dirname(filePath);
 }