+++ CHANGELOG.md
@@ -7,7 +7,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- bug fix - untagging component with dependencies was showing an error saying the component is depends on itself
+
+## [0.12.6-dev.4] - 2018-02-19
+
+- save bit dev-dependencies components inside devDependencies section of package.json
+- fix bit cc to clear module's cache
+- fix resolving .gitignore files
+- when adding existing files to component in bitmap don't create duplications
+
+## [0.12.6-dev.3] - 2018-02-15
+
+- support `bit install` of specific ids
 - support peerDependencies (in the model for now)
+- [#729](https://github.com/teambit/bit/issues/729) fix bit cc to clear module cache
+- [#769](https://github.com/teambit/bit/issues/769) - prevent duplicate ids in bitmap when adding existing files
 
 ## [0.12.6-dev.2] - 2018-02-13
 
@@ -16,6 +30,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `bit status` shows a list of staged versions in 'staged components' section
 - introduced a new command `bit untag` for reverting un-exported tags.
 - init local scope inside .git
+- support vue files
+- [#736](https://github.com/teambit/bit/issues/736) - .gitignore is blocking everything
 
 ## [0.12.5] - 2018-02-06